### 2022-04-21 #4334 @editorialbot
[PRE REVIEW]: sam2lca: Lowest Common Ancestor for SAM/BAM/CRAM alignment files
**Submitting author:** @maxibor (<a href="http://orcid.org/0000-0001-9140-7559">Maxime Borry</a>)
**Repository:** https://github.com/maxibor/sam2lca
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0b
