+++ paper/paper.bib
@@ -31,6 +31,7 @@
 @article{herbig2017malt,
   title={MALT: Fast alignment and analysis of metagenomic DNA sequence data applied to the Tyrolean Iceman},
   author={Herbig, Alexander and Maixner, Frank and Bos, Kirsten I and Zink, Albert and Krause, Johannes and Huson, Daniel H},
+  journal={BioRxiv},
   year={2017},
   publisher={Cold Spring Harbor Laboratory},
   doi={10.1101/050559}
@@ -82,12 +83,11 @@
 @article{bbmap,
 title = {BBMap: A Fast, Accurate, Splice-Aware Aligner},
 author = {Bushnell, Brian},
-abstractNote = {Alignment of reads is one of the primary computational tasks in bioinformatics. Of paramount importance to resequencing, alignment is also crucial to other areas - quality control, scaffolding, string-graph assembly, homology detection, assembly evaluation, error-correction, expression quantification, and even as a tool to evaluate other tools. An optimal aligner would greatly improve virtually any sequencing process, but optimal alignment is prohibitively expensive for gigabases of data. Here, we will present BBMap [1], a fast splice-aware aligner for short and long reads. We will demonstrate that BBMap has superior speed, sensitivity, and specificity to alternative high-throughput aligners bowtie2 [2], bwa [3], smalt, [4] GSNAP [5], and BLASR [6].},
 doi = {},
 url = {https://www.osti.gov/biblio/1241166}, 
 journal = {},
-number = ,
-volume = ,
+number = {},
+volume = {},
 place = {United States},
 year = {2014},
 month = {3}
@@ -133,10 +133,3 @@ month = {3}
   journal = {GitHub repository},
   url = {https://github.com/pysam-developers/pysam}
 }
-
-
-
-
-
-
-
