### 2022-04-29 eddca1d
add edit distance bam write
