### 2022-05-04 maxibor/sam2lca#10 @marouenbg
sam2lca JOSS review #1
Hi @maxibor ,

I reviewed sam2lca as a submission to JOSS (https://github.com/openjournals/joss-reviews/issues/4360). I think it is an excellent piece of software, very well documented, and the application note was very well written and clear. It was easy to appreciate the amount of effort you put in have a neat repo, unit tests, and complete set of docs.

I had a few minor comments:
In the paper:
- L26 should be in order to place instead of in order place
- Also, I understand that this is an IO interface to taxopy, if that is the case, the gap would be to improve the usability of taxopy along with the general research question of mapping ambiguous reads.

In the docs:
- It could be worth mentioning the 4 steps of sam2lca in the docs as well.
-I attempted to run the tutorial, which is very thorough and complete, real-world use case. I suggest listing all the prerequisites and where to install them in the beggining: fastp and bowtie2. Samtools was not mentioned at all and popped in one of the command lines.
- I had an error in this part
![image](https://user-images.githubusercontent.com/7914548/166704554-7da69447-a955-49b0-a01f-e12ea323c7e2.png)

In the repo:
- I clone the repo and tried running the tests but I had this error (I am on MacOs)
![image](https://user-images.githubusercontent.com/7914548/166701194-662a46d8-935a-49b7-bbeb-41bd6eeadf56.png)
- In readme, the link to SAM/BAM/CRAM is not working
![image](https://user-images.githubusercontent.com/7914548/166701379-6bcb6a07-9339-4840-a6c6-5e2d92cf1e02.png)
- I suggest adding how to run units tests in the installation section `For development purposes, from the dev branch`

I am pretty sure that these errors are due to missing software/proper configuration on my end. Let me know if you need me to develop on them further.

Marouen
