### 2022-05-05 9c57aec
update paper after https://github.com/maxibor/sam2lca/issues/10
