+++ paper/paper.md
@@ -44,7 +44,7 @@ The rapidly decreasing cost of massively parallel short-read DNA sequencing tech
 
 In a typical shotgun metagenomics approach, after the DNA of an ecological community has been sequenced, it is compared to a genetic reference database of organisms with known taxonomy. Even though the number of DNA sequences and genomes in reference databases is constantly growing, there are still instances where a query sequence will not have a direct match in a reference database, and it will instead weakly align to one or more distantly related reference organisms. Furthermore, when analyzing short DNA sequences, a query DNA sequence will often match equally well to more than one reference organism, posing a challenge for its taxonomic assignation.
 
-One solution to this problem is to apply a lowest common ancestor algorithm (LCA) (\autoref{fig:fig1}) during taxonomic profiling in order place such ambiguous assignments higher in a taxonomic tree, where they can be more confidently assigned. This idea was first implemented for metagenomics with the MEGAN program [@huson2007megan].
+One solution to this problem is to apply a lowest common ancestor algorithm (LCA) (\autoref{fig:fig1}) during taxonomic profiling in order to place such ambiguous assignments higher in a taxonomic tree, where they can be more confidently assigned. This idea was first implemented for metagenomics with the MEGAN program [@huson2007megan].
 
 Many programs have since been developed to perform LCA during taxonomic profiling. For example, MALT [@herbig2017malt] and MetaPhlan [@segata2012metagenomic] perform LCA and taxonomic profiling after DNA sequence alignment, while other programs, such as Kraken2 [@kraken2] and Centrifuge [@kim2016centrifuge], are alignment-free methods that apply LCA after k-mer matching. While combining the steps of database matching and LCA into one program can be useful, it also limits user choice with respect to the selection of different alignment or k-mer matching programs.
 
