### 2022-05-05 2b2eda7
explicitly list additional tools needed for tutorial
