+++ docs/source/contributing.md
@@ -4,21 +4,23 @@ We welcome any contributions !
 
 To further develop sam2lca, or add documentation, please read further:
 
-## Clone the sam2lca repository, and checkout the dev branch
+## 1. Clone the sam2lca repository, and checkout the dev branch
 
 ```bash
 git clone git@github.com:maxibor/sam2lca.git
 git checkout dev
 ```
 
-## Install and activate the development environment
+## 2. Install and activate the sam2lca conda environment
 
 ```bash
 conda env create -f environment.yml
 conda activate sam2lca
 ```
 
-## Install sam2lca with pip in editable mode
+> For all the following steps, we strongly suggest you to work in the `sam2lca` conda environment
+
+## 3. Install sam2lca with pip in editable mode
 
 ```bash
 pip install -e .
