### 2022-05-05 893c3d2
update readme with @marouenbg suggestions
