### 2022-05-05 maxibor/sam2lca#10 @maxibor
Hey @marouenbg ,
First of all, thanks for taking the time to review sam2lca !

> L26 should be in order to place instead of in order place

Thanks for spotting it, it's now fixed with [9c57aec](https://github.com/maxibor/sam2lca/commit/9c57aecf1eee54b700aaa6a96f9ee7a7c2e8c663)

>Also, I understand that this is an IO interface to taxopy, if that is the case, the gap would be to improve the usability of taxopy along with the general research question of mapping ambiguous reads.

We view as a bit more than a IO interface to taxopy, as there are extra steps required after parsing the SAM files (like converting the accesion numbers to taxids using RocksDB for example) before being able to use taxopy. Futhermore, our target user will mainly use sam2lca through the command line, which doesn't expose taxopy directly.

>It could be worth mentioning the 4 steps of sam2lca in the docs as well.
-I attempted to run the tutorial, which is very thorough and complete, real-world use case. I suggest listing all the prerequisites and where to install them in the beggining: fastp and bowtie2. Samtools was not mentioned at all and popped in one of the command lines.

In the tutorial, we handle the dependancy requirement with a conda environment, which is described at the beginning: https://github.com/maxibor/sam2lca/blob/edaa34ac5d0db125a847b5b404e5334a95ebc148/docs/source/tutorial.md#installing-all-tools-for-this-tutorial

But I nevertheless listed them explictely in the tutorial with [2b2eda7](https://github.com/maxibor/sam2lca/commit/2b2eda78e9aa83d6328f415666a773c118a7a9e0)

From the look of your path, it looks like the `sam2lca_tutorial` environment is maybe not activated ?

> I clone the repo and tried running the tests but I had this error (I am on MacOs)

Here again, we suggest working in a conda environment, and it looks like the environment is not activated on your machine https://github.com/maxibor/sam2lca/blob/edaa34ac5d0db125a847b5b404e5334a95ebc148/docs/source/contributing.md#install-and-activate-the-development-environment
I nevertheless now explicitly added that we suggest to work in the conda environment with [3a0404a](https://github.com/maxibor/sam2lca/commit/3a0404a3b793f9524ef0daa39b67a945662078e0)

> In readme, the link to SAM/BAM/CRAM is not working

Thanks for spotting it ! It is now fixed with [893c3d2](https://github.com/maxibor/sam2lca/commit/893c3d231257736ba9d82721cff93f9317688bfa)

> I suggest adding how to run units tests in the installation section For development purposes, from the dev branch

I added it as suggested with [893c3d2](https://github.com/maxibor/sam2lca/commit/893c3d231257736ba9d82721cff93f9317688bfa)
