### 2022-05-05 dec4193
add 18s pr2 to plant markers db
