+++ tests/test_acc2tax_db.py
@@ -1,6 +1,6 @@
 import os
-from sam2lca.data import acc2tax_default
 from sam2lca.acc2tax import get_mapping
+from sam2lca.utils import get_json
 from sam2lca.rocksdb import OPTS_read
 import rocksdb
 from shutil import rmtree
@@ -14,12 +14,14 @@ maptype = "test"
 
 
 def test_create_db():
-    get_mapping(map_config=acc2tax_default, maptype=maptype, dbdir=db_dir)
+    config = get_json("https://raw.githubusercontent.com/maxibor/sam2lca/master/data/acc2tax.json")
+    get_mapping(map_config=config, maptype=maptype, dbdir=db_dir)
 
 
 def test_acc2tax_db_query():
+    config = get_json("https://raw.githubusercontent.com/maxibor/sam2lca/master/data/acc2tax.json")
     db = rocksdb.DB(
-        os.path.join(db_dir, acc2tax_default["map_db"][maptype]),
+        os.path.join(db_dir, config["map_db"][maptype]),
         OPTS_read,
         read_only=True,
     )
