+++ sam2lca/data/gtdb_taxdump/__init__.py
@@ -1,14 +1,0 @@
-gtdb_taxdump = {
-    "r207": {
-        "url": {
-            "names": "https://raw.githubusercontent.com/maxibor/sam2lca/taxopy/sam2lca/data/gtdb_taxdump/r207/names.dmp.gz",
-            "nodes": "https://raw.githubusercontent.com/maxibor/sam2lca/taxopy/sam2lca/data/gtdb_taxdump/r207/nodes.dmp.gz",
-            "merged": "https://raw.githubusercontent.com/maxibor/sam2lca/taxopy/sam2lca/data/gtdb_taxdump/r207/merged.dmp.gz",
-        },
-        "md5": {
-            "names": "https://raw.githubusercontent.com/maxibor/sam2lca/taxopy/sam2lca/data/gtdb_taxdump/r207/names.dmp.gz.md5",
-            "nodes": "https://raw.githubusercontent.com/maxibor/sam2lca/taxopy/sam2lca/data/gtdb_taxdump/r207/nodes.dmp.gz.md5",
-            "merged": "https://raw.githubusercontent.com/maxibor/sam2lca/taxopy/sam2lca/data/gtdb_taxdump/r207/merged.dmp.gz.md5",
-        },
-    }
-}
