+++ tests/test_acc2tax_db.py
@@ -7,7 +7,6 @@ from shutil import rmtree
 import pytest
 
 test_dir = os.path.join(os.path.dirname(__file__))
-data_dir = os.path.join(test_dir, "data")
 db_dir = os.path.join(test_dir, "test_acc2tax_dbdir")
 os.makedirs(db_dir, exist_ok=True)
 maptype = "test"
