+++ tests/test_update_list_db.py
@@ -4,7 +4,6 @@ from sam2lca.main import list_available_db
 import pytest
 
 test_dir = os.path.join(os.path.dirname(__file__))
-data_dir = os.path.join(test_dir, "data")
 db_dir = os.path.join(test_dir, "test_update_dbdir")
 os.makedirs(db_dir, exist_ok=True)
 
