### 2022-05-12 9769535
handle more unknown taxid errors
