### 2022-05-16 #4360 @marouenbg
@jmschrei I finished reviewing sam2Ica, there were only minor comments really, the authors did a fantastic job presenting the problem and their approach, recognizing other software tools they based their analysis on, and particularly did a great job packaging the tool and the docs. Overall, excellent submission!
