+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to [sam2lca](https://github.com/maxibor/sam2lca) will be doc
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.0.0beta
+## 1.0.0
 
 ### Added
 
