### 2022-05-27 ee14a45
update conda recipe
