+++ conda/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "sam2lca" %}
-{% set version = "1.0.0b0" %}
+{% set version = "1.0.0" %}
 
 package:
   name: {{ name|lower }}
@@ -7,7 +7,7 @@ package:
 
 source:
   url: https://pypi.io/packages/source/{{ name[0] }}/{{ name }}/sam2lca-{{ version }}.tar.gz
-  sha256: 6d91c363cd0aa5d13c2e3491da4241109ece3948cb038d5a621e8263cf584f60
+  sha256: 4d41b16443a84611c67a5892a32c97ace90efb4cab1abfe0645409fef739cc88
 
 build:
   entry_points:
