### 2022-05-27 #4360 @maxibor
Many thanks for the constructive reviews @fasnicar and @marouenbg , and thanks a lot for the editing @jmschrei !!
I've now made a new release (1.0.0) with a Zenodo DOI: `10.5281/zenodo.6585307`
