### 2022-06-01 #4360 @danielskatz
Sorry for the delay in processing this - I missed it in the handover between AEiCs that happens weekly, but I've now proofread the paper and found nothing that needs to change, so once the zenodo metadata is updated, we can proceed to final acceptance and publication
